+++ lib/main.dart
@@ -1,12 +1,19 @@
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:app_home_demo/view/root.dart';
 
-List todoList = [
-  ["科目A ", "期限切れの課題 ", DateTime.now().subtract(const Duration(days: 1)), "S", 3]
-];
+// List todoList = [
+//   ["科目A ", "期限切れの課題 ", DateTime.now().subtract(const Duration(days: 1)), "S", 3]
+// ];
+
+Future<void> main2() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+}
 
 void main() {
+  main2();
   runApp(
     const ProviderScope(
       child: MyApp(),
